requirements.txt
streamlit==1.38.0
streamlit-player==0.1.5
pytube==15.0.0
youtube-search-python==1.6.6
httpx==0.27.2
yt-dlp>=2024.10.22,<2026
ffmpeg-python==0.2.0
imageio-ffmpeg==0.4.9
faster-whisper==1.0.2
chromadb==0.5.5
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
transformers==4.44.2
timm==1.0.9
easyocr==1.7.1
opencv-python-headless==4.9.0.80
matplotlib==3.8.4
plotly==5.24.1
Pillow==10.4.0
safetensors==0.4.3
accelerate==0.34.2
sentence-transformers==3.1.0
numpy==1.26.4
pydantic==2.9.2
reportlab==4.2.2
python-docx==1.1.2
wordcloud==1.9.3
